+++ src/Libraries/SmartStore.Services/Tasks/IScheduleTaskService.cs
@@ -50,12 +50,6 @@ namespace SmartStore.Services.Tasks
 		/// <summary>
 		/// Ensures that a task is not marked as running
 		/// </summary>
-		/// <param name="type">Task type</param>
-		void EnsureTaskIsNotRunning(string type);
-
-		/// <summary>
-		/// Ensures that a task is not marked as running
-		/// </summary>
 		/// <param name="taskId">Task identifier</param>
 		void EnsureTaskIsNotRunning(int taskId);
     }